requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
Pillow
jinja2
python-multipart
